MIS Assignment 3 -MurrietaSmith
This is my first project which demonstrates my abilities in SQL. In this project I had to create tables in SQL, use primary and foreign keys to show relationships between tables, add data into tables, and retrieve data from those tables. These are fundamental to using SQL and syntax is important to understand as well as how to fix it when it's incorrect. The drop table functions at the top demonstrate the use of functions to troubleshoot code that I've written. Overall this project shows that I am able to use SQL to build and use relations for data.